Prevent memory churn described in #208
Accessing `stateFlow.value` is not guaranteed to always return the same instance even when the underlying value is unchanged. This  stateFlow behavior was incorrectly causing infinite `objectWillChange` events due to the equality test always reporting `false` even when the new value was, in fact, equal to the previous.

The fix is to go through [`isEqual(to:)`](https://developer.apple.com/documentation/foundation/nsarray/1411770-isequal) to compare the value _contents_ when determine equality before firing off an `objectWillChange` event.